restructure tests & code
