--- sample-1-output.txt
@@ -1,30 +1,0 @@
-+-------+-------------------+--------+------------+-----+-------+------------+----------------------------+--------+
-|   ID  |        Name       | Gender |  Birthday  | Age | Alive |   Death    |           Child            | Spouse |
-+-------+-------------------+--------+------------+-----+-------+------------+----------------------------+--------+
-| @I10@ | Mordechai Thomson |   M    | 1962-05-06 |  58 |  True |     NA     |             NA             |  @I9@  |
-| @I11@ |    Knox Wilder    |   M    | 2005-09-10 |  15 |  True |     NA     |             NA             |   NA   |
-| @I12@ |    Nash Wilder    |   M    | 2006-10-11 |  14 | False | 2006-10-12 |             NA             |   NA   |
-| @I13@ |   Cohen Thomson   |   M    |     NA     |  NA |  True |     NA     |     ['@I10@', '@I15@']     | @I14@  |
-| @I14@ |  Addyson Thomson  |   F    |     NA     |  NA |  True |     NA     |     ['@I10@', '@I15@']     | @I13@  |
-| @I15@ |  Coraline Thomson |   F    | 1964-12-07 |  56 |  True |     NA     |             NA             |   NA   |
-|  @I1@ |   Manuel Wilder   |   M    | 2000-01-01 |  21 |  True |     NA     |             NA             |  @I2@  |
-|  @I2@ |   Davina Hardin   |   F    | 2001-03-03 |  19 |  True |     NA     |             NA             |  @I1@  |
-|  @I3@ |   Cassius Hardin  |   M    | 1981-05-05 |  39 |  True |     NA     |          ['@I2@']          |  @I4@  |
-|  @I4@ |     Ayla Burch    |   F    | 1980-04-04 |  40 |  True |     NA     |          ['@I2@']          |  @I3@  |
-|  @I5@ |   Andres Wilder   |   M    | 1983-08-08 |  37 |  True |     NA     | ['@I1@', '@I11@', '@I12@'] |  @I7@  |
-|  @I6@ |   Kaitlyn Rogers  |   F    | 1982-07-07 |  38 |  True |     NA     |          ['@I1@']          |   NA   |
-|  @I7@ |    Emmy Holder    |   F    | 1984-09-09 |  36 |  True |     NA     |     ['@I11@', '@I12@']     |  @I5@  |
-|  @I8@ |    Jamir Wilder   |   M    | 1961-03-04 |  59 | False | 2010-04-05 |          ['@I5@']          |  @I9@  |
-|  @I9@ |   Jimena Sanchez  |   F    | 1960-01-02 |  61 |  True |     NA     |          ['@I5@']          | @I10@  |
-+-------+-------------------+--------+------------+-----+-------+------------+----------------------------+--------+
-+------+------------+------------+------------+-------------------+---------+-----------------+--------------------+
-|  ID  |  Married   |  Divorced  | Husband ID |    Husband Name   | Wife ID |    Wife Name    |      Children      |
-+------+------------+------------+------------+-------------------+---------+-----------------+--------------------+
-| @F1@ | 2019-02-02 |     NA     |    @I1@    |   Manuel Wilder   |   @I2@  |  Davina Hardin  |         NA         |
-| @F2@ | 1998-11-11 | 2000-11-12 |    @I5@    |   Andres Wilder   |   @I6@  |  Kaitlyn Rogers |      ['@I1@']      |
-| @F3@ | 1997-12-12 |     NA     |    @I3@    |   Cassius Hardin  |   @I4@  |    Ayla Burch   |      ['@I2@']      |
-| @F4@ | 2001-10-10 |     NA     |    @I5@    |   Andres Wilder   |   @I7@  |   Emmy Holder   | ['@I11@', '@I12@'] |
-| @F5@ | 1978-02-03 |     NA     |    @I8@    |    Jamir Wilder   |   @I9@  |  Jimena Sanchez |      ['@I5@']      |
-| @F6@ | 2011-06-07 |     NA     |   @I10@    | Mordechai Thomson |   @I9@  |  Jimena Sanchez |         NA         |
-| @F7@ |     NA     |     NA     |   @I13@    |   Cohen Thomson   |  @I14@  | Addyson Thomson | ['@I10@', '@I15@'] |
-+------+------------+------------+------------+-------------------+---------+-----------------+--------------------+
