+++ validation.py
@@ -1,0 +1,29 @@
+import utils as utils
+
+'''
+  Implements US08
+  
+  Returns a list pairs:
+  [(id1, r1), (id2, r2), ...]
+  
+  Where the first item in the pair is the id of an
+  invalid family and the second item is the reason.
+  
+  Note that a given family id may appear more than
+  once if multiple invalid reasons are found.
+'''
+def validate_birth_marriage_order(fams, indis):
+  ret_data = []
+  
+  for fid in fams:
+    if fams[fid]['MARR'] is not None:
+      marriage = utils.parse_date(fams[fid]['MARR'])
+      
+      for cid in fams[fid]['CHIL']:
+        if indis[cid]['BIRT'] is not None:
+          birth = utils.parse_date(indis[cid]['BIRT'])
+          if birth < marriage:
+            ret_data.append((fid, f'Child id={cid} has birthdate before marriage'))
+  
+  return ret_data
+
